fix: Ensure `UsbReadFailedError` can be pickled
This change ensures that the base `Exception` class is aware of the
`usb_error` arg which gets passed to the subclass, rather than
intercepting it. By doing so, we leverage the built-in `__reduce__()`
behaviour of `Exception` so that `UsbReadFailedError` can be pickled and
unpickled without blowing up.

This change causes a minor change to the representation of these
exceptions, However stringification remains unchanged. Previously:
```python
>>> e = adb_shell.exceptions.UsbReadFailedError("foo", usb1.USBError("bar"))
>>> str(e)
'foo: Unknown error [bar]'
>>> repr(e)
"UsbReadFailedError('foo')"
```

and with this change:
```python
>>> e = adb_shell.exceptions.UsbReadFailedError("foo", usb1.USBError("bar"))
>>> str(e)
'foo: Unknown error [bar]'
>>> repr(e)
"UsbReadFailedError('foo', USBErrorTimeout())"
```