fix: Use idempotent `USBContext` in `USBTransport`
This avoids an issue with our lack of lifecycle management for the
`USBContext` objects created while finding ADB devices over USB. In
0.4.0 and earlier versions, a context object would be created each time
a caller attempted to find devices using the `USBTransport` and it would
be inadvertently leaked. The GC may have been able to clean these
objects up later once the devices using the context object were also
released.

The change in e7d8b59 attempted to clean up a `DeprecationWarning` being
issues by `python-libusb1` which was telling us to use the `USBContext`
object as a context manager. Doing this had the effect of introducing
explicit (but wrong) lifecycle management of the `USBContext` object and
its underlying C objects from `libusb1.so`. Since the `USBContext` would
now be released once the device iterator was finalised, it was likely
that a device returned to code calling into `adb_shell` would attempt to
use (through several layers of abstraction) a `usb1` device object which
had had its underlying context object released and probably freed.

This change attempts to mitigate the object leakage by maintaining an
idempotent `USBContext` object in our `USBTransport` class object, to
ensure that as long as the class exists, instances of it can continue to
use that context object.